Updates based on review feedback
