+++ lib/etl/redshift/client.rb
@@ -6,6 +6,7 @@ require 'odbc'
 require 'mixins/cached_logger'
 require 'pathname'
 require 'fileutils'
+require_relative 'stl_load_error'
 
 module ETL::Redshift
   # when the odbc driver is setup in chef this is the driver's name
@@ -496,18 +497,4 @@ SQL
   class SkipRow
   end
 
-  class RedshiftSTLLoadError < StandardError
-    attr_accessor :error_row, :detail_rows, :error_s3_file, :local_error_file
-    def initialize(error_row, detail_rows)
-      @error_row = error_row
-      @detail_rows = detail_rows
-    end
-    def message
-      row = {}
-      @detail_rows.each do |dr|
-        row[dr[:colname].strip] = dr[:value].to_s.strip
-      end
-      "STL Load error: Reason: '#{@error_row[:err_reason].strip}', LineNumber: #{@error_row[:line_number]}, Position: #{@error_row[:position]}, Rawline '#{error_row[:raw_line].strip}', \nparsed row: '#{row.to_s}'"
-    end
-  end
 end
